ENH: Documented tractography computation and cleaned code
